# src/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dextora API",
    "version": "0.1.0",
    "description": "REST API for the Dextora backend (Bun + Express). Use **Authorize** with a JWT from `POST /api/v1/auth/login` or `POST /api/v1/auth/register`."
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "variables": {
        "port": { "default": "3000" }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT from login or register (`token` field in the response)."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "issues": { "type": "object", "additionalProperties": true }
            },
            "required": ["message"]
          }
        },
        "required": ["error"]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid JWT",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      },
...
      "ValidationError": {
        "description": "Request validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ApiError" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/parse/pdf-export": {
      "post": {
        "tags": ["Parse"],
        "summary": "Parse PDF to nested chapters/topics/atoms JSON",
        "description": "Synchronous export: extracts text, detects chapters and topics, pairs paragraphs into atoms, runs batched parallel Gemini classification (when configured), attaches generation prompts (quiz, game, simulation, video, and `prompts.illustrationImage` per atom/topic/chapter for external image models). TTS and Gemini-generated artifacts (quiz, games, topic/chapter content) are queued automatically; poll GET /api/v1/parse/export/{exportId}/generated. Response includes optional `generation.queued`. SuperTTS HTTP when SUPERTTS_HTTP_URL is set; else Gemini TTS when GEMINI_TTS_MODEL is set. Workers: `JOB_QUEUE_DRIVER=redis` with `REDIS_URL` (e.g. redis://127.0.0.1:6379) and `bun run worker`; jobs use BullMQ retries (3 attempts, exponential backoff). HTML games: `PARSE_EXPORT_HTML_VERIFY_MODE` (relaxed|strict) and `PARSE_EXPORT_HTML_MAX_BYTES`; `PARSE_EXPORT_ATOM_INTERNAL_CONCURRENCY` parallelises sub-generations inside one atom/topic job. Optional `PUBLIC_API_BASE_URL` makes TTS `audioUrl` absolute. Query: classifyConcurrency (default 12), classifyBatchSize (default 16, max 48), ttsConcurrency (reserved for workers), ttsMaxAtoms (default 50, 0 skips TTS jobs). PDF up to ~120MB.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "classifyConcurrency",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "default": 12 }
          },
          {
            "name": "classifyBatchSize",
            "in": "query",
            "description": "Atoms per Gemini classification request (fewer round trips when higher). Capped at 48.",
            "schema": { "type": "integer", "minimum": 1, "maximum": 48, "default": 16 }
          },
          {
            "name": "ttsConcurrency",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "ttsMaxAtoms",
            "in": "query",
            "description": "Max atoms to synthesize TTS for (highest importance first). Use 0 to skip TTS.",
            "schema": { "type": "integer", "minimum": 0, "default": 50 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nested parse result",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/parse/export/{exportId}/status": {
      "get": {
        "tags": ["Parse"],
        "summary": "Parse-export generation progress",
        "description": "Read-only: counts completed artifact jobs vs total for an export owned by the caller.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Progress snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exportId": { "type": "string", "format": "uuid" },
                    "complete": { "type": "boolean" },
                    "ready": { "type": "boolean" },
                    "progress": {
                      "type": "object",
                      "properties": {
                        "done": { "type": "integer" },
                        "total": { "type": "integer" },
                        "failedCells": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
    "/api/v1/parse/export/{exportId}/generated": {
      "get": {
        "tags": ["Parse"],
        "summary": "Fetch async parse-export artifacts",
        "description": "Read-only: returns per-atom, per-topic, and per-chapter generated payloads when present (quiz JSON, HTML, TTS `audioUrl`, etc.). When `PUBLIC_API_BASE_URL` is set at generation time, `audioUrl` is an absolute URL; otherwise it stays a same-origin path under `/api/v1/files/audio`. Nested `prompts.illustrationImage` strings are ready-to-send image model prompts. Poll until `complete` is true.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Artifacts and progress",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
  }
}